Спасибо за комментарии!
Комментарии, полученные через  GitHub
1. Модели Вынесены в Отдельные файлы - группа Models (первоначально сделал все в одном файле, т.к. в ТЗ требование - «Весь код написан в одном файле MovieQuizViewController.swift»
2.     let padding = UIEdgeInsets(top: 5, left: 20, bottom: 5, right: 20)  - удалено
3.  Форматирование массива - поправил
4. Отступы во всех файлах выровнены
5. «Куда-то пропал пробел между 2 скобками в конце строчки)» - добавил

Комментарии через раздел «Ревью» в Практикум
1. Нужно исправить верстку (пункт 3.1) шрифты, к сожалению, неправильно импортированы. При открытии Storyboard файла показывается ошибка Font “YSDisplay-Bold” does not exist on this system, substituting the system font instead. Saving the document will lose the original font reference., указывающая на это. Советую посмотреть в сторону правильного импорта шрифтов и вообще ресурсов в проект

Ошибка не воспроизводится. Для контроля заново импортировал шрифты (в т.ч. переустановил шрифт на мой Mac) и переназначил на объектах. В «Build Phases»  .ttf файлы находятся в «Copy Bundle Resources». Через print Family Fonts выводятся в консоль. Дополнительно, в инфоплсте добавил явно расширение файла .ttf. В симуляторе  применяются.

2. В консоль пишется ошибка Failed to set (keyPath) user defined inspected property on (UIView): [<UIView 0x158f28a50> setValue:forUndefinedKey:]: this class is not key value coding-compliant for the key keyPath.. - исправлено.
3. «Обращу внимание, что пользователь может случайно прожать два раза кнопку ответа или даже много и быстро понажимать на кнопки» - добавил логику отключения кнопок (дополнительно проверил через isHidden))  )
