+++ MovieQuiz/Presentation/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Альберт Хайдаров on 08.08.2023.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let imageOfFilm: String
+    let questionText: String
+    let correctAnswer: Bool
+}
