+++ MovieQuiz/Presentation/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Альберт Хайдаров on 08.08.2023.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let alertTitle: String
+    let resultText: String
+    let alertButtonText: String
+}
