+++ MovieQuiz/Presentation/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Альберт Хайдаров on 08.08.2023.
+//
+
+import Foundation
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
