+++ MovieQuiz/Resources/Info.plist
@@ -2,13 +2,15 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>UIAppFonts</key>
+	<array>
+		<string>YS Display-Bold.ttf</string>
+		<string>YS Display-Medium.ttf</string>
+	</array>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIAppFonts</key>
-		<array>
-			<string>YS Display-Bold</string>
-			<string>YS Display-Medium</string>
-		</array>
+		<array/>
 		<key>UIApplicationSupportsMultipleScenes</key>
 		<false/>
 		<key>UISceneConfigurations</key>
